feat: Introduce BigInteger for RFID and add README refractor: code cleanup

+++ app/src/main/java/com/example/iot_project/models/Client.kt
@@ -1,8 +1,10 @@
 package com.example.iot_project.models
 
+import java.math.BigInteger
+
 data class Client(
     val id: Int,
-    val rfid: Int,
+    val rfid: BigInteger,
     val name: String,
     val lastname: String,
     val email: String,
